+++ src/DataContainer/OfferFeedAttributeContainer.php
@@ -40,7 +40,7 @@ class OfferFeedAttributeContainer extends Backend
      */
     public function getFieldOptions(): array
     {
-        return ['text', 'select', 'picker', 'fileTree', 'listWizard'];
+        return ['text', 'textarea', 'select', 'picker', 'fileTree', 'listWizard'];
     }
 
     /**