+++ src/EventListener/LoadDataContainerListener.php
@@ -17,6 +17,7 @@ namespace WEM\OffersBundle\EventListener;
 use Psr\Log\LoggerInterface;
 use WEM\UtilsBundle\Classes\StringUtil;
 use WEM\OffersBundle\Model\OfferFeedAttribute;
+use WEM\UtilsBundle\Classes\StringUtil;
 
 class LoadDataContainerListener
 {
@@ -47,7 +48,7 @@ class LoadDataContainerListener
         }
     }
 
-    protected function parseDcaAttribute($row): array
+    protected function parseDcaAttribute(array $row): array
     {
         // Generic data
         $data = [
@@ -66,8 +67,8 @@ class LoadDataContainerListener
 
         // Maxlength settings
         if ($row['maxlength']) {
-            $data['eval']['maxlength'] = (int) $row['maxlength'];
-            $data['sql']['length'] = (int) $row['maxlength'];
+            $data['eval']['maxlength'] = (int)$row['maxlength'];
+            $data['sql']['length'] = (int)$row['maxlength'];
         }
 
         // Available for alerts settings
@@ -87,9 +88,24 @@ class LoadDataContainerListener
             $data['eval']['mandatory'] = true;
         }
 
+        // rte settings
+        if ($row['rte']) {
+            $data['eval']['rte'] = $row['rte'];
+        }
+
+        // Class settings
+        if ($row['explanation']) {
+            $data['explanation'] = $row['explanation'];
+        }
+
         // Class settings
         if ($row['class']) {
             $data['eval']['tl_class'] = $row['class'];
+        }
+
+        // Allow helpwizard
+        if ($row['helpwizard']) {
+            $data['eval']['helpwizard'] = true;
         }
 
         switch ($row['type']) {
@@ -108,6 +124,16 @@ class LoadDataContainerListener
                     $data['default'] = '';
                     $data['sql']['default'] = '';
                 }
+
+                break;
+
+            case 'textarea':
+                // Allow HTML settings
+                if ($row['allowHtml']) {
+                    $data['eval']['allowHtml'] = true;
+                }
+
+                $data['sql'] = 'mediumtext NULL';
 
                 break;