+++ src/Module/ModuleOffersList.php
@@ -27,6 +27,7 @@ use WEM\OffersBundle\Model\Offer as OfferModel;
 use WEM\UtilsBundle\Classes\StringUtil;
 use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
+use WEM\OffersBundle\Model\Offer;
 
 /**
  * Front end module "offers list".
@@ -125,42 +126,7 @@ class ModuleOffersList extends ModuleOffers
         }
 
         // Catch Ajax requets
-        if (Input::post('TL_AJAX') && (int) $this->id === (int) Input::post('module')) {
-            try {
-                switch (Input::post('action')) {
-                    case 'seeDetails':
-                        if (!Input::post('offer')) {
-                            throw new \Exception(sprintf($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['argumentMissing'], 'offer'));
-                        }
-
-                        $objItem = OfferModel::findByPk(Input::post('offer'));
-
-                        $this->offer_template = 'offer_details';// TODO : InsertTag
-                        System::getContainer()->get('contao.insert_tag')->replace($this->parseOffer($objItem));
-                        exit;
-
-                    case 'apply':
-                        if (!Input::post('offer')) {
-                            throw new \Exception(sprintf($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['argumentMissing'], 'offer'));
-                        }
-
-                        // Put the offer in session
-                        $objSession->set('wem_offer', Input::post('offer'));
-                        System::getContainer()->get('contao.insert_tag')->replace($this->getApplicationForm((int)Input::post('offer')));
-                        exit;
-
-                    default:
-                        throw new \Exception(sprintf($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['unknownRequest'], Input::post('action')));
-                }
-            } catch (\Exception $e) {
-                $arrResponse = ['status' => 'error', 'msg' => $e->getResponse(), 'trace' => $e->getTrace()];
-            }
-
-            // Add Request Token to JSON answer and return
-            $arrResponse['rt'] = $this->csrfTokenManager->getToken($this->csrfTokenName)->getValue();
-            echo json_encode($arrResponse);
-            exit;
-        }
+        $this->catchAjaxRequests();
 
         if ($this->offer_applicationForm
             && '' !== $objSession->get('wem_offer')
@@ -188,7 +154,7 @@ class ModuleOffersList extends ModuleOffers
             $this->limit = $this->numberOfItems;
         }
 
-        $this->Template->articles = [];
+        $this->Template->items = [];
         $this->Template->empty = $GLOBALS['TL_LANG']['WEM']['OFFERS']['empty'];
 
         // assets
@@ -203,11 +169,35 @@ class ModuleOffersList extends ModuleOffers
         $this->config = ['pid' => $this->offer_feeds, 'published' => 1];
 
         // Retrieve filters
-        $this->buildFilters();
-        $this->Template->filters = $this->filters;
+        if (!empty($_GET) || !empty($_POST)) {
+            foreach ($_GET as $f => $v) {
+                if (false === strpos($f, 'offer_filter_')) {
+                    continue;
+                }
+
+                if (Input::get($f)) {
+                    $this->config[str_replace('offer_filter_', '', $f)] = Input::get($f);
+                }
+            }
+
+            foreach ($_POST as $f => $v) {
+                if (false === strpos($f, 'offer_filter_')) {
+                    continue;
+                }
+
+                if (Input::post($f)) {
+                    $this->config[str_replace('offer_filter_', '', $f)] = Input::post($f);
+                }
+            }
+        }
+
+        // Retrieve filters
+        if ($this->offer_addFilters) {
+            $this->Template->filters = $this->getFrontendModule($this->offer_filters_module);
+        }
 
         // Get the total number of items
-        $intTotal = OfferModel::countItems($this->config);
+        $intTotal = Offer::countItems($this->config);
 
         if ($intTotal < 1) {
             return;
@@ -246,188 +236,21 @@ class ModuleOffersList extends ModuleOffers
             $this->Template->pagination = $objPagination->generate("\n  ");
         }
 
-        $objArticles = OfferModel::findItems($this->config, ($this->limit ?: 0), ($this->offset ?: 0));
+        $objItems = Offer::findItems($this->config, ($this->limit ?: 0), ($this->offset ?: 0));
 
         // Add the articles
-        if ($objArticles instanceof \Contao\Model\Collection) {
-            $this->Template->articles = $this->parseOffers($objArticles);
+        if ($objItems instanceof \Contao\Model\Collection) {
+            $this->Template->items = $this->parseOffers($objItems);
         }
 
         $this->Template->moduleId = $this->id;
 
         // Catch auto_item
         if (Input::get('auto_item')) {
-            $objOffer = OfferModel::findItems(['code' => Input::get('auto_item')], 1);
+            $objOffer = Offer::findItems(['code' => Input::get('auto_item')], 1);
 
             $this->Template->openModalOnLoad = true;
             $this->Template->offerId = $objOffer->first()->id;
-        }
-    }
-
-    /**
-     * Parse and return an application form for a job.
-     *
-     * @param int $intId       [Job ID]
-     * @param string $strTemplate [Template name]
-     */
-    protected function getApplicationForm(int $intId, string $strTemplate = 'offer_apply'): string
-    {
-        if (!$this->offer_applicationForm) {
-            return '';
-        }
-
-        $strForm = $this->getForm($this->offer_applicationForm);
-
-        $objItem = OfferModel::findByPk($intId);
-
-        $objTemplate = new FrontendTemplate($strTemplate);
-        $objTemplate->id = $objItem->id;
-        $objTemplate->code = $objItem->code;
-        $objTemplate->title = $objItem->title;
-        $objTemplate->recipient = $GLOBALS['TL_ADMIN_EMAIL'];
-        $objTemplate->time = time();
-        $objTemplate->token = $this->csrfTokenManager->getToken($this->csrfTokenName)->getValue();
-        $objTemplate->form = $strForm;
-
-        return $objTemplate->parse();
-    }
-
-    /**
-     * Retrieve list filters.
-     *
-     * @throws \Exception
-     */
-    protected function buildFilters(): void
-    {
-        if (!$this->offer_addFilters) {
-            exit();
-        }
-
-        // Retrieve and format dropdowns filters
-        $filters = StringUtil::deserialize($this->offer_filters);
-        if (\is_array($filters) && $filters !== []) {
-            foreach ($filters as $f) {
-                $field = $GLOBALS['TL_DCA']['tl_wem_offer']['fields'][$f];
-
-                $filter = [
-                    'type' => $field['inputType'],
-                    'name' => $field['eval']['multiple'] ? $f.'[]' : $f,
-                    'label' => $field['label'][0] ?: $GLOBALS['TL_LANG']['tl_wem_offer'][$f][0],
-                    'value' => Input::get($f) ?: '',
-                    'options' => [],
-                    'multiple' => (bool)$field['eval']['multiple'],
-                ];
-
-                switch ($field['inputType']) {
-                    case 'select':
-                        if (\is_array($field['options_callback'])) {
-                            $strClass = $field['options_callback'][0];
-                            $strMethod = $field['options_callback'][1];
-
-                            $this->import($strClass);
-                            $options = $this->$strClass->$strMethod($this);
-                        } elseif (\is_callable($field['options_callback'])) {
-                            $options = $field['options_callback']($this);
-                        } elseif (\is_array($field['options'])) {
-                            $options = $field['options'];
-                        }
-
-                        foreach ($options as $value => $label) {
-                            if (\is_array($label)) {
-                                foreach ($label as $subValue => $subLabel) {
-                                    $filter['options'][$value]['options'][] = [
-                                        'value' => $subValue,
-                                        'label' => $subLabel,
-                                        'selected' => (null !== Input::get($f) && (Input::get($f) === $subValue || (\is_array(Input::get($f)) && \in_array($subValue, Input::get($f), true)))),
-                                    ];
-                                }
-                            } else {
-                                $filter['options'][] = [
-                                    'value' => $value,
-                                    'label' => $label,
-                                    'selected' => (null !== Input::get($f) && (Input::get($f) === $value || (\is_array(Input::get($f)) && \in_array($value, Input::get($f), true)))),
-                                ];
-                            }
-                        }
-
-                        break;
-
-                    case 'listWizard':
-                        $objOptions = OfferModel::findItemsGroupByOneField($f);
-
-                        if ($objOptions) {
-                            $filter['type'] = 'select';
-                            if ($filter['multiple']) {
-                                $filter['name'] .= '[]';
-                            }
-
-                            while ($objOptions->next()) {
-                                if (!$objOptions->{$f}) {
-                                    continue;
-                                }
-
-                                $subOptions = StringUtil::deserialize($objOptions->{$f});
-                                foreach ($subOptions as $subOption) {
-                                    $filter['options'][$subOption] = [
-                                        'value' => $subOption,
-                                        'label' => $subOption,
-                                        'selected' => $filter['multiple']
-                                            ? (null !== Input::get($f) && \in_array($subOption, Input::get($f ?? []), true))
-                                            : (null !== Input::get($f) && Input::get($f) === $subOption),
-                                    ];
-                                }
-                            }
-                        }
-
-                        break;
-
-                    case 'text':
-                    default:
-                        $objOptions = OfferModel::findItemsGroupByOneField($f);
-
-                        if ($objOptions && 0 < $objOptions->count()) {
-                            $filter['type'] = 'select';
-                            while ($objOptions->next()) {
-                                if (!$objOptions->{$f}) {
-                                    continue;
-                                }
-
-                                $filter['options'][] = [
-                                    'value' => $objOptions->{$f},
-                                    'label' => $objOptions->{$f},
-                                    'selected' => (null !== Input::get($f) && Input::get($f) === $objOptions->{$f}),
-                                ];
-                            }
-                        }
-
-                        break;
-                }
-
-                if ('select' === $filter['type'] && 1 >= \count($filter['options'])) {
-                    continue;
-                }
-
-                if (null !== Input::get($f) && '' !== Input::get($f)) {
-                    $this->config[$f] = Input::get($f);
-                }
-
-                $this->filters[] = $filter;
-            }
-        }
-
-        // Add fulltext search if asked
-        if ($this->offer_addSearch) {
-            $this->filters[] = [
-                'type' => 'text',
-                'name' => 'search',
-                'label' => $GLOBALS['TL_LANG']['WEM']['OFFERS']['search'],
-                'placeholder' => $GLOBALS['TL_LANG']['WEM']['OFFERS']['searchPlaceholder'],
-                'value' => Input::get('search') ?: '',
-            ];
-
-            if ('' !== Input::get('search') && null !== Input::get('search')) {
-                $this->config['search'] = StringUtil::formatKeywords(Input::get('search'));
-            }
         }
     }
 }