+++ src/Resources/contao/dca/tl_wem_offer.php
@@ -115,6 +115,8 @@ $GLOBALS['TL_DCA']['tl_wem_offer'] = [
             'sql' => "int(10) unsigned NOT NULL default '0'",
         ],
         'pid' => [
+            'foreignKey' => 'tl_wem_offer_feed.title',
+            'relation' => ['type' => 'belongsTo', 'load' => 'eager'],
             'sql' => "int(10) unsigned NOT NULL default '0'",
         ],
         'createdAt' => [