+++ src/Resources/contao/dca/tl_wem_offer_feed.php
@@ -79,6 +79,7 @@ $GLOBALS['TL_DCA']['tl_wem_offer_feed'] = [
     'palettes' => [
         'default' => '
             {title_legend},title,alias;
+            {config_legend},jumpTo;
             {attributes_legend},attributes;
             {alert_legend},ncEmailAlert,tplOfferAlert
         ',
@@ -114,6 +115,15 @@ $GLOBALS['TL_DCA']['tl_wem_offer_feed'] = [
                 [OfferFeedContainer::class, 'generateAlias'],
             ],
             'sql' => "varchar(255) BINARY NOT NULL default ''",
+        ],
+
+        'jumpTo' => [
+            'exclude' => true,
+            'inputType' => 'pageTree',
+            'foreignKey' => 'tl_page.title',
+            'eval' => array('fieldType'=>'radio'),
+            'sql' => "int(10) unsigned NOT NULL default 0",
+            'relation' => array('type'=>'hasOne', 'load'=>'lazy')
         ],
 
         'attributes' => [