+++ src/Resources/contao/dca/tl_wem_offer_feed_attribute.php
@@ -80,6 +80,7 @@ $GLOBALS['TL_DCA']['tl_wem_offer_feed_attribute'] = [
     // Subpalettes
     'subpalettes' => [
         'type_text' => 'value,isFilter,isAlertCondition',
+        'type_textarea' => 'allowHtml,helpwizard,rte,explanation',
         'type_select' => 'options,multiple,chosen,isFilter,isAlertCondition',
         'type_picker' => 'fkey',
         'type_fileTree' => 'multiple,filesOnly,fieldType,extensions',
@@ -95,6 +96,8 @@ $GLOBALS['TL_DCA']['tl_wem_offer_feed_attribute'] = [
             'sql' => "int(10) unsigned NOT NULL default '0'",
         ],
         'pid' => [
+            'foreignKey' => 'tl_wem_offer_feed.title',
+            'relation' => ['type' => 'belongsTo', 'load' => 'eager'],
             'sql' => "int(10) unsigned NOT NULL default '0'",
         ],
         'createdAt' => [
@@ -130,6 +133,13 @@ $GLOBALS['TL_DCA']['tl_wem_offer_feed_attribute'] = [
             'exclude' => true,
             'inputType' => 'text',
             'eval' => ['decodeEntities' => true, 'maxlength' => 255, 'tl_class' => 'w50'],
+            'sql' => "varchar(255) NOT NULL default ''",
+        ],
+        'rte' => [
+            'exclude' => true,
+            'search' => true,
+            'inputType' => 'text',
+            'eval' => ['maxlength' => 255, 'tl_class' => 'w50'],
             'sql' => "varchar(255) NOT NULL default ''",
         ],
         'options' => [
@@ -187,6 +197,12 @@ $GLOBALS['TL_DCA']['tl_wem_offer_feed_attribute'] = [
             'eval' => ['maxlength' => 255, 'tl_class' => 'w50 cbx'],
             'sql' => "char(1) NOT NULL default ''",
         ],
+        'helpwizard' => [
+            'exclude' => true,
+            'inputType' => 'checkbox',
+            'eval' => ['maxlength' => 255, 'tl_class' => 'w50 cbx'],
+            'sql' => "char(1) NOT NULL default ''",
+        ],
         'mandatory' => [
             'exclude' => true,
             'filter' => true,
@@ -223,6 +239,18 @@ $GLOBALS['TL_DCA']['tl_wem_offer_feed_attribute'] = [
             'sql' => ['name' => 'insertType', 'type' => 'string', 'length' => 128, 'default' => 'POSITION_APPEND'],
         ],
         'class' => [
+            'exclude' => true,
+            'inputType' => 'text',
+            'eval' => ['maxlength' => 255, 'tl_class' => 'w50'],
+            'sql' => "varchar(255) NOT NULL default ''",
+        ],
+        'rte' => [
+            'exclude' => true,
+            'inputType' => 'text',
+            'eval' => ['maxlength' => 255, 'tl_class' => 'w50'],
+            'sql' => "varchar(255) NOT NULL default ''",
+        ],
+        'explanation' => [
             'exclude' => true,
             'inputType' => 'text',
             'eval' => ['maxlength' => 255, 'tl_class' => 'w50'],